+++ src/id_translation/transform/_impl/bitmask.py
@@ -38,6 +38,8 @@ class BitmaskTransformer(Transformer[IdType]):
         joiner: A string used to join bitmask flag labels.
         overrides: A dict ``{id: translation}``. Use to add or override the translation source.
         force_decomposition: If ``True``, ignore composite values in the translation source.
+        force_real_translations: If ``True``, convert :class:`id_translation.offline.MagicDict` instances to plain
+            ``dict`` using the :attr:`.MagicDict.real` attribute. This will prevent translations such as ``''``.
 
     Examples:
         Basic usage.
@@ -57,6 +59,19 @@ class BitmaskTransformer(Transformer[IdType]):
 
         Note that ``0='NOT_SET'`` was translated even though it's not in the ``data``, and that ``8='0b1000'`` was
         replaced by ``'OVERFLOW!'``, as per the overrides specified for the transformer.
+
+        Implication of setting ``force_real_translations=False``.
+
+        >>> btr = BitmaskTransformer(force_real_translations=False)
+        >>> tra = Translator({"bitmasks": data}, transformers={"bitmasks": btr})
+        >>> tra.translate((5, 6), names="bitmasks", max_fails=0.0)
+        ('1:name-of-1 & 4:name-of-4', '<Failed: id=2> & 4:name-of-4')
+
+        The translation "succeeded", even though ``max_fails=0.0`` and ``6 = '<Failed: id=2> & 4:name-of-4'`` was only a
+        partial success. This would've raised :class:`an error <.TooManyFailedTranslationsError>` if
+        `force_real_translations` was not set. The transformer adds :attr:`~.MagicDict.real` mappings for all composite
+        IDs, so the :class:`.Translator` won't detect any issues when using :meth:`.MagicDict.real_contains` to verify
+        the results.
     """
 
     def __init__(
@@ -65,6 +80,7 @@ class BitmaskTransformer(Transformer[IdType]):
         *,
         overrides: Mapping[IdType, str] | None = None,
         force_decomposition: bool = False,
+        force_real_translations: bool = True,
     ) -> None:
         self._joiner = joiner
         self._force = force_decomposition
@@ -72,6 +88,7 @@ class BitmaskTransformer(Transformer[IdType]):
             # TOML keys must be strings, so we use a record format.
             overrides = self._from_toml_records(overrides)
         self._overrides = overrides or {}
+        self._force_real_translations = force_real_translations
 
     @classmethod
     def update_ids(cls, ids: set[IdType]) -> None:
@@ -108,6 +125,11 @@ class BitmaskTransformer(Transformer[IdType]):
             return
 
     def _create_composite_translation(self, bits: list[IdType], *, translations: Mapping[IdType, str]) -> str:
+        from id_translation.offline import MagicDict
+
+        if self._force_real_translations and isinstance(translations, MagicDict):
+            translations = translations.real
+
         return self._joiner.join(translations[idx] for idx in bits)
 
     def __repr__(self) -> str: